Test now compiling (added dependencies, fixed typing errors in code, etc etc etc

+++ src/main/java/com/google/songgame/data/GuessChecker.java
@@ -2,16 +2,20 @@ package com.google.songgame.data;
 
 import com.google.appengine.api.datastore.Entity;
 import com.google.appengine.api.datastore.EmbeddedEntity;
+import com.google.cloud.datastore.testing.LocalDatastoreHelper;
 
 /**
- * Contains various helper methods revolving around checking a users guess including checking whether a user has already correctly guessed, whether a guess is correct, and updating the game state if a user has guessed correctly
+ * Contains various helper methods revolving around checking a users guess including checking
+ * whether a user has already correctly guessed, whether a guess is correct, and updating the game
+ * state if a user has guessed correctly
  */
 public final class GuessChecker {
-  private static final int POINTS_PER_ROUND = 100;
+  private static final long POINTS_PER_ROUND = 100;
   /**
    * Checks if the user has already guessed correctly.
-   * 
-   * <p>Checks for the current round if the user has already guessed correctly and returns true or false
+   *
+   * <p>Checks for the current round if the user has already guessed correctly and returns true or
+   * false
    */
   public static boolean hasUserPreviouslyGuessedCorrect(Entity user, Entity game) {
     String userId = (String) user.getProperty("userId");
@@ -24,8 +28,9 @@ public final class GuessChecker {
 
   /**
    * Checks if the given guess is correct.
-   * 
-   * <p>Checks if the guess is correct and matches the video title of the current round, ignoring case
+   *
+   * <p>Checks if the guess is correct and matches the video title of the current round, ignoring
+   * case
    */
   public static boolean isCorrectGuess(String guess, Entity game) {
     EmbeddedEntity currentRound = (EmbeddedEntity) game.getProperty("currentRound");
@@ -37,8 +42,9 @@ public final class GuessChecker {
 
   /**
    * Updates the game state to indicate the user has correctly guessed
-   * 
-   * <p>Updates for the current round that the user has guessed correctly. Updates in the userGuessStatuses map stored in the currentRound entity of the game.
+   *
+   * <p>Updates for the current round that the user has guessed correctly. Updates in the
+   * userGuessStatuses map stored in the currentRound entity of the game.
    */
   public static Entity markUserGuessedCorrectly(Entity user, Entity game) {
     String userId = (String) user.getProperty("userId");
@@ -49,10 +55,10 @@ public final class GuessChecker {
     game.setProperty("currentRound", currentRound);
     return game;
   }
-  
+
   /**
    * Updates the game state by increasing the users points
-   * 
+   *
    * <p>Updates for the given user the point total. Updates in the userPoints map.
    */
   public static Entity assignUserPoints(Entity user, Entity game) {
